replace old local autotest with compile stage one
