+++ test/local_test_autosplit.py
@@ -7,33 +7,21 @@ import pippy.fx
 import pippy.ModelSplit
 
 import torch
+import torch.distributed as dist
 import torch.autograd.profiler_legacy
 from pippy import run_pippy
+from pippy.compile import compile_stage
 from pippy.IR import MultiUseParameterConfig, Pipe
-from pippy.PipelineDriver import (
-    PipelineDriver1F1B,
-    PipelineDriverBase,
-    PipelineDriverFillDrain,
-    PipelineDriverInterleaved1F1B,
-)
 
 PROFILING_ENABLED = True
 CHECK_NUMERIC_EQUIVALENCE = True
 
-schedules = {
-    "FillDrain": PipelineDriverFillDrain,
-    "1F1B": PipelineDriver1F1B,
-    "Interleaved1F1B": PipelineDriverInterleaved1F1B,
-}
-
 pippy.fx.Tracer.proxy_buffer_attributes = True
-
-MULTI_USE_PARAM_CONFIG = MultiUseParameterConfig.TRANSMIT
 
 
 # Example model definition
 d_hid = 512
-bs = 503
+bs = 503  # batch_size
 
 
 class ExampleCode(torch.nn.Module):
@@ -42,7 +30,7 @@ class ExampleCode(torch.nn.Module):
         self.mm_param = torch.nn.Parameter(torch.randn(d_hid, d_hid))
         self.mm_param2 = torch.nn.Parameter(torch.randn(d_hid, d_hid))
         self.lin = torch.nn.Linear(d_hid, d_hid)
-        self.register_buffer("buffer", torch.randn(bs + 100, d_hid))
+        self.register_buffer("buffer", torch.randn(bs, d_hid))
 
     def forward(self, x):
         x = torch.mm(x, self.mm_param)
@@ -55,7 +43,8 @@ class ExampleCode(torch.nn.Module):
         x = torch.mm(x, self.mm_param2)
         x = self.lin(x)
         x = torch.relu(x)
-        return {"out": x}
+
+        return x
 
 
 def inspect_split_module(
@@ -78,84 +67,113 @@ def inspect_split_module(
         print(submod)
 
 
-# Common function to run pipeline with input and check equivalence
-def run_pipe_driver(ec_pipe, args):
-    nstages = len(list(ec_pipe.split_gm.children()))
-
-    pipe_driver: PipelineDriverBase = schedules[args.schedule](
-        ec_pipe,
-        nstages,
-        args.world_size,
-        _debug_mask_minibatches=True,
-        _record_mem_dumps=bool(args.record_mem_dumps),
-        checkpoint=bool(args.checkpoint),
-    )
-
-    # # Warm up and correctness runs
+# Common function to run pipeline stage with input and check equivalence
+def run_compile_stage(mod: torch.nn.Module, stage: Pipe, args):
+    nstages = len(list(stage.split_gm.children()))
     ec_input = torch.randn(bs, d_hid, device=args.device)
-    out = pipe_driver(ec_input)
-    ref_out = ec_pipe(ec_input)
+
+    # Warm up and correctness runs
+    if args.rank == 0:
+        stage(ec_input)
+    elif args.rank == args.world_size - 1:
+        out = stage()
+    else:
+        stage()
+
+    print(f"Rank {args.rank} completes")
 
     # run with different chunk size to exercise microbatch and scheduling components
-    pipe_driver.chunks = 1
-    pipe_driver(ec_input)
-    pipe_driver.chunks = 100
-    pipe_driver(ec_input)
+    # stage.chunks = 1
+    # stage(ec_input)
+    # stage.chunks = 100
+    # stage(ec_input)
 
-    if CHECK_NUMERIC_EQUIVALENCE:
-        torch.testing.assert_close(out["out"], ref_out["out"])
-        print(
-            f'equivalence test passed {torch.sum(out["out"])} ref {torch.sum(ref_out["out"])}'
-        )
+    # check numeric equivalence in the last rank
+    if args.rank == args.world_size - 1:
+        if CHECK_NUMERIC_EQUIVALENCE:
+            ref_out = mod(ec_input)
+            torch.testing.assert_close(out, ref_out)
+            print(
+                f'equivalence test passed {torch.sum(out["out"])} ref {torch.sum(ref_out["out"])}'
+            )
 
-    # # Profiling runs
+    # Profiling runs
     with torch.autograd.profiler_legacy.profile(
         enabled=PROFILING_ENABLED
     ) as prof:
-        pipe_driver.chunks = nstages
-        out = pipe_driver(ec_input)
-        ref_out = ec_pipe(ec_input)
-        print(
-            f'profiling run completed {torch.sum(out["out"])} ref {torch.sum(ref_out["out"])}'
-        )
+        stage.chunks = nstages
+
+        if args.rank == 0:
+            stage(ec_input)
+        elif args.rank == args.world_size - 1:
+            out = stage(ec_input)
+        else:
+            stage()
+
+        ref_out = mod(ec_input)
+
+        if args.rank == args.world_size - 1:
+            print(f"#############{out.shape}################")
+            print(
+                f'profiling run completed {torch.sum(out)} ref {torch.sum(ref_out)}'  #TODO: bug in square brackets: I get `profiling run completed [nan/inf] ref 188975536.0`
+            )
     if PROFILING_ENABLED:
         prof.export_chrome_trace(
             f"{os.path.splitext(os.path.basename(__file__))[0]}.json"
         )
 
+    dist.barrier()
 
-def test_split_on_size_threshold(_, args):
+
+def test_split_on_size_threshold(args):
     ec = ExampleCode()
     ec.to(args.device)
 
     # Auto-split based on size threshold
     threshold = 300000
     split_policy = pippy.ModelSplit.split_on_size_threshold(threshold)
-    ec_pipe = Pipe.from_tracing(
-        ec, MULTI_USE_PARAM_CONFIG, split_policy=split_policy
+
+    ec_input = torch.randn(bs, d_hid, device=args.device)
+    target = torch.randn(bs, d_hid, device=args.device)
+
+    stage = compile_stage(  # default multi_use_param_spec to REPLICATE'ing param(if used in multiple stages) across stages instead of TRANSMI'ing it
+        ec,
+        args.rank,
+        args.world_size,
+        args.chunks,
+        args.device,
+        None,
+        [ec_input,],
+    #     split_policy=split_policy
     )
 
-    inspect_split_module(ec_pipe, expected_stages=5)
+    # inspect_split_module(stage, expected_stages=5)
+    # run_compile_stage(ec, stage, args)
 
-    run_pipe_driver(ec_pipe, args)
 
-
-def test_split_into_equal_size(_, args):
+def test_split_into_equal_size(args):
     ec = ExampleCode()
     ec.to(args.device)
-    ec_input = torch.randn(bs, d_hid, device=args.device)
-    ec(ec_input)
 
     # Auto-split based on given number of stages
     nstages = 5
     split_policy = pippy.ModelSplit.split_into_equal_size(nstages)
-    ec_pipe = Pipe.from_tracing(
-        ec, MULTI_USE_PARAM_CONFIG, split_policy=split_policy
+
+    ec_input = torch.randn(bs, d_hid, device=args.device)
+    target = torch.randn(bs, d_hid, device=args.device)
+    stage = compile_stage(  # default multi_use_param_spec to REPLICATE'ing param(if used in multiple stages) across stages instead of TRANSMI'ing it
+        ec,
+        args.rank,
+        args.world_size,
+        args.chunks,
+        args.device,
+        None,
+        [ec_input, ],
+        split_policy=split_policy
     )
 
-    inspect_split_module(ec_pipe, expected_stages=nstages)
-
-    run_pipe_driver(ec_pipe, args)
+    # inspect_split_module(stage, expected_stages=nstages)
+    run_compile_stage(ec, stage, args)
 
 
 def main(args=None):
@@ -163,22 +181,17 @@ def main(args=None):
     parser.add_argument(
         "--world_size", type=int, default=int(os.getenv("WORLD_SIZE", 5))
     )
-    parser.add_argument("--rank", type=int, default=int(os.getenv("RANK", -1)))
+    parser.add_argument(
+        "--rank", type=int, default=int(os.getenv("RANK", -1))
+    )
+    parser.add_argument(
+        "--chunks", type=int, default=5,
+    )
     parser.add_argument(
         "--master_addr", type=str, default=os.getenv("MASTER_ADDR", "localhost")
     )
     parser.add_argument(
         "--master_port", type=str, default=os.getenv("MASTER_PORT", "29500")
-    )
-    parser.add_argument(
-        "-s",
-        "--schedule",
-        type=str,
-        default=list(schedules.keys())[0],
-        choices=schedules.keys(),
-    )
-    parser.add_argument(
-        "--replicate", type=int, default=int(os.getenv("REPLICATE", "0"))
     )
     parser.add_argument(
         "--cuda", type=int, default=int(torch.cuda.is_available())
@@ -189,18 +202,19 @@ def main(args=None):
     parser.add_argument("--checkpoint", type=int, default=0, choices=[0, 1])
     args = parser.parse_args(args)
 
-    # Interleaved 1F1B uses less ranks than number of stages
-    if args.schedule == "Interleaved1F1B":
-        args.world_size = 2
+    args.device = torch.device(f"cuda:{args.rank % torch.cuda.device_count()}") \
+                    if args.cuda else torch.device("cpu")
 
-    global MULTI_USE_PARAM_CONFIG
-    if args.replicate:
-        MULTI_USE_PARAM_CONFIG = MultiUseParameterConfig.REPLICATE
-    print(f"REPLICATE config: {args.replicate} -> {MULTI_USE_PARAM_CONFIG}")
-    print("Using schedule:", args.schedule)
+    # init process group
+    backend = "nccl" if args.cuda else "gloo"
+    dist.init_process_group(
+        backend=backend,
+        rank=args.rank,
+        world_size=args.world_size,
+    )
 
-    run_pippy(test_split_on_size_threshold, args)
-    run_pippy(test_split_into_equal_size, args)
+    test_split_into_equal_size(args)
+    # test_split_on_size_threshold(args)
 
 
 if __name__ == "__main__":
